requirements.txt
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-SocketIO==4.3.2
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.0
python-engineio==3.14.2
python-socketio==4.6.1
six==1.16.0
Werkzeug==2.0.1
